requirements.txt
certifi==2020.12.5
click==7.1.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.5.1.48
Pillow==8.1.1
Werkzeug==1.0.1
wincertstore==0.2
